requirements.txt
selenium==3.14.1
cx_Oracle
pytest
pytest-repeat==0.8.0
pytest-rerunfailures==8.0
pytest-ordering==0.6
allure-pytest
